Create review controller and move complete action to it
