+++ app/controllers/candidate_interface/personal_details/base_controller.rb
@@ -39,36 +39,6 @@ module CandidateInterface
         end
       end
 
-      def show
-        @application_form = current_application
-        @personal_details_form = PersonalDetailsForm.build_from_application(current_application)
-        @nationalities_form = NationalitiesForm.build_from_application(current_application)
-        @languages_form = LanguagesForm.build_from_application(current_application)
-        @personal_details_review = PersonalDetailsReviewPresenter.new(
-          personal_details_form: @personal_details_form,
-          nationalities_form: @nationalities_form,
-          languages_form: @languages_form,
-        )
-      end
-
-      def complete
-        binding.pry
-        if PersonalDetailsForm.new(current_application).valid? &&
-          NationalitiesForm.new(current_application).valid? &&
-          LanguagesForm.new(current_application).valid?
-
-          current_application.update!(application_form_params)
-
-          redirect_to candidate_interface_application_form_path
-        else
-          @personal_details_form = PersonalDetailsForm.build_from_application(current_application)
-          @nationalities_form = NationalitiesForm.build_from_application(current_application)
-          @languages_form = LanguagesForm.build_from_application(current_application)
-
-          render :show
-        end
-      end
-
     private
 
       def personal_details_params
@@ -77,11 +47,6 @@ module CandidateInterface
           :"date_of_birth(3i)", :"date_of_birth(2i)", :"date_of_birth(1i)"
         )
           .transform_keys { |key| dob_field_to_attribute(key) }
-          .transform_values(&:strip)
-      end
-
-      def application_form_params
-        params.require(:application_form).permit(:personal_details_completed)
           .transform_values(&:strip)
       end
