+++ app/views/candidate_interface/personal_details/review/show.html.erb
