+++ config/routes.rb
@@ -72,8 +72,8 @@ Rails.application.routes.draw do
         post '/nationalities' => 'personal_details/nationalities#update'
         get '/languages' => 'personal_details/languages#edit', as: :languages
         post '/languages' => 'personal_details/languages#update'
-        get '/review' => 'personal_details/base#show', as: :personal_details_show
-        post '/' => 'personal_details/base#complete', as: :personal_details_complete
+        get '/review' => 'personal_details/review#show', as: :personal_details_show
+        post '/review' => 'personal_details/review#complete', as: :personal_details_complete
       end
 
       scope '/personal-statement' do
